backend/requirements.txt
flask>=2.0.0
python-dotenv>=0.19.0
docker>=6.0.0
gitpython>=3.1.30
